// src/types.ts
// types.ts
export interface ContactInfo {
    email: string;
    linkedIn: string;
    github: string;
  }
  
  export interface Skill {
    name: string;
  }
  
  export interface Experience {
    title: string;
    company: string;
    date: string;
    description: string;
  }
  
  export interface Education {
    degree: string;
    institution: string;
    graduationYear: string;
  }
  
  export interface Project {
    title: string;
    description: string;
    link: string;
  }
  